@@ -9,33 +9,33 @@ namespace Zaabee.NewtonsoftJson
 {
     public static partial class NewtonsoftJsonHelper
     {
-        public static async Task<MemoryStream> PackAsync<T>(T t, JsonSerializerSettings settings = null,
+        public static Task<MemoryStream> PackAsync<T>(T t, JsonSerializerSettings settings = null,
             Encoding encoding = null) =>
             t is null
-                ? new MemoryStream()
-                : await NewtonsoftJsonSerializer.PackAsync(t, settings ?? DefaultSettings,
+                ? Task.FromResult(new MemoryStream())
+                : NewtonsoftJsonSerializer.PackAsync(t, settings ?? DefaultSettings,
                     encoding ?? DefaultEncoding);
 
-        public static async Task<MemoryStream> PackAsync(Type type, object obj, JsonSerializerSettings settings = null,
+        public static Task<MemoryStream> PackAsync(Type type, object obj, JsonSerializerSettings settings = null,
             Encoding encoding = null) =>
             obj is null
-                ? new MemoryStream()
-                : await NewtonsoftJsonSerializer.PackAsync(type, obj, settings ?? DefaultSettings,
+                ? Task.FromResult(new MemoryStream())
+                : NewtonsoftJsonSerializer.PackAsync(type, obj, settings ?? DefaultSettings,
                     encoding ?? DefaultEncoding);
 
-        public static async Task PackAsync<T>(T t, Stream stream, JsonSerializerSettings settings = null,
+        public static Task PackAsync<T>(T t, Stream stream, JsonSerializerSettings settings = null,
             Encoding encoding = null)
         {
-            if (t is null || stream is null) return;
-            await NewtonsoftJsonSerializer.PackAsync(t, stream, settings ?? DefaultSettings,
+            if (t is null || stream is null) return Task.CompletedTask;
+            return NewtonsoftJsonSerializer.PackAsync(t, stream, settings ?? DefaultSettings,
                 encoding ?? DefaultEncoding);
         }
 
-        public static async Task PackAsync(Type type, object obj, Stream stream, JsonSerializerSettings settings = null,
+        public static Task PackAsync(Type type, object obj, Stream stream, JsonSerializerSettings settings = null,
             Encoding encoding = null)
         {
-            if (obj is null || stream is null) return;
-            await NewtonsoftJsonSerializer.PackAsync(type, obj, stream, settings ?? DefaultSettings,
+            if (obj is null || stream is null) return Task.CompletedTask;
+            return NewtonsoftJsonSerializer.PackAsync(type, obj, stream, settings ?? DefaultSettings,
                 encoding ?? DefaultEncoding);
         }
 
@@ -47,11 +47,11 @@ namespace Zaabee.NewtonsoftJson
         /// <param name="encoding"></param>
         /// <typeparam name="T"></typeparam>
         /// <returns></returns>
-        public static async Task<T> UnpackAsync<T>(Stream stream, JsonSerializerSettings settings = null,
+        public static Task<T> UnpackAsync<T>(Stream stream, JsonSerializerSettings settings = null,
             Encoding encoding = null) =>
             stream is null
-                ? (T) typeof(T).GetDefaultValue()
-                : await NewtonsoftJsonSerializer.UnpackAsync<T>(stream, settings ?? DefaultSettings,
+                ? Task.FromResult((T) typeof(T).GetDefaultValue())
+                : NewtonsoftJsonSerializer.UnpackAsync<T>(stream, settings ?? DefaultSettings,
                     encoding ?? DefaultEncoding);
 
         /// <summary>
@@ -62,11 +62,11 @@ namespace Zaabee.NewtonsoftJson
         /// <param name="settings"></param>
         /// <param name="encoding"></param>
         /// <returns></returns>
-        public static async Task<object> UnpackAsync(Type type, Stream stream, JsonSerializerSettings settings = null,
+        public static Task<object> UnpackAsync(Type type, Stream stream, JsonSerializerSettings settings = null,
             Encoding encoding = null) =>
             stream is null
-                ? type.GetDefaultValue()
-                : await NewtonsoftJsonSerializer.UnpackAsync(type, stream, settings ?? DefaultSettings,
+                ? Task.FromResult(type.GetDefaultValue())
+                : NewtonsoftJsonSerializer.UnpackAsync(type, stream, settings ?? DefaultSettings,
                     encoding ?? DefaultEncoding);
     }
 }
