@@ -6,21 +6,20 @@ namespace Zaabee.MsgPack
 {
     public static partial class MsgPackHelper
     {
-        public static async Task<MemoryStream> PackAsync<T>(T t)
+        public static Task<MemoryStream> PackAsync<T>(T t) =>
+            t is null
+                ? Task.FromResult(new MemoryStream())
+                : MsgPackSerializer.PackAsync(t);
+
+        public static Task PackAsync<T>(T t, Stream stream)
         {
-            if (t is null) return new MemoryStream();
-            return await MsgPackSerializer.PackAsync(t);
+            if (t is null || stream is null) return Task.CompletedTask;
+            return MsgPackSerializer.PackAsync(t, stream);
         }
 
-        public static async Task PackAsync<T>(T t, Stream stream)
-        {
-            if (t is null || stream is null) return;
-            await MsgPackSerializer.PackAsync(t, stream);
-        }
-
-        public static async Task<T> UnpackAsync<T>(Stream stream) =>
+        public static Task<T> UnpackAsync<T>(Stream stream) =>
             stream is null
-                ? (T) typeof(T).GetDefaultValue()
-                : await MsgPackSerializer.UnpackAsync<T>(stream);
+                ? Task.FromResult((T) typeof(T).GetDefaultValue())
+                : MsgPackSerializer.UnpackAsync<T>(stream);
     }
 }
