@@ -9,10 +9,10 @@ namespace Zaabee.Xml
         /// Serializes the specified object and writes the XML document to a file using the specified XmlWriter.
         /// </summary>
         /// <param name="xmlWriter"></param>
-        /// <param name="obj"></param>
+        /// <param name="t"></param>
         /// <typeparam name="T"></typeparam>
-        public static void Serialize<T>(XmlWriter xmlWriter, object obj) =>
-            Serialize(typeof(T), xmlWriter, obj);
+        public static void Serialize<T>(XmlWriter xmlWriter, T t) =>
+            Serialize(typeof(T), xmlWriter, t);
 
         /// <summary>
         /// Serializes the specified object and writes the XML document to a file using the specified XmlWriter.
